+++ packages/react-integration/demo-app-ts/src/components/demos/LabelDemo/LabelDemo.tsx
@@ -1,5 +1,5 @@
 import { Label } from '@patternfly/react-core';
-import React, { Component } from 'react';
+import { Component } from 'react';
 import InfoCircleIcon from '@patternfly/react-icons/dist/esm/icons/info-circle-icon';
 import { Link } from 'react-router-dom';
 
@@ -10,7 +10,7 @@ export class LabelDemo extends Component {
 
   render() {
     return (
-      <React.Fragment>
+      <>
         <Label id="no-tooltip">Grey</Label> <Label icon={<InfoCircleIcon />}>Grey icon</Label>{' '}
         <Label onClose={() => {}}>Grey removeable</Label>{' '}
         <Label icon={<InfoCircleIcon />} onClose={() => {}}>
@@ -20,7 +20,7 @@ export class LabelDemo extends Component {
         <Label href="#" onClose={() => {}}>
           Grey link removeable
         </Label>
-        <span style={{ width: "150px", display: "inline-block"}}>
+        <span style={{ width: '150px', display: 'inline-block' }}>
           <Label id="tooltip" variant="outline">
             Very very very very long label text that should be truncated without textMaxWidth
           </Label>
@@ -80,7 +80,7 @@ export class LabelDemo extends Component {
         >
           Edit
         </Label>
-      </React.Fragment>
+      </>
     );
   }
 }