+++ packages/react-integration/demo-app-ts/src/components/demos/ModalDemo/ModalDemo.tsx
@@ -285,14 +285,14 @@ export class ModalDemo extends React.Component<React.HTMLProps<HTMLDivElement>, 
     const { isCustomHeaderFooterModalOpen } = this.state;
 
     const header = (
-      <React.Fragment>
+      <>
         <Title id="customHeaderTitle" headingLevel="h1" size={TitleSizes['2xl']}>
           Custom Modal Header/Footer
         </Title>
         <p id="customHeaderDescription" className={spacing.ptSm}>
           Allows for custom content in the header and/or footer by passing components.
         </p>
-      </React.Fragment>
+      </>
     );
 
     const footer = (
@@ -482,7 +482,7 @@ export class ModalDemo extends React.Component<React.HTMLProps<HTMLDivElement>, 
     };
 
     return (
-      <React.Fragment>
+      <>
         <div id="tabstop-test" tabIndex={0} style={{ display: 'flex', flexDirection: 'row', flexWrap: 'wrap' }}>
           <Button style={buttonStyle} variant="primary" onClick={this.handleModalToggle} id="showDefaultModalButton">
             Show Modal
@@ -575,7 +575,7 @@ export class ModalDemo extends React.Component<React.HTMLProps<HTMLDivElement>, 
         {this.renderModalWithAlertVariant()}
         {this.renderHelpModal()}
         {this.renderCustomFocusModal()}
-      </React.Fragment>
+      </>
     );
   }
 }