+++ packages/react-integration/demo-app-ts/src/components/demos/ModalNextDemo/ModalNextDemo.tsx
@@ -97,14 +97,14 @@ export class ModalNextDemo extends React.Component<React.HTMLProps<HTMLDivElemen
   handleModalCustomEscapeToggle = (_event?: any, customEscapePressed?: boolean) => {
     this.setState(({ isModalCustomEscapeOpen }) => ({
       isModalCustomEscapeOpen: !isModalCustomEscapeOpen,
-      customEscapePressed: customEscapePressed ?? false 
+      customEscapePressed: customEscapePressed ?? false
     }));
   };
 
   handleModalAlertVariantToggle = (_event?: any, customEscapePressed?: boolean) => {
     this.setState(({ isModalAlertVariantOpen }) => ({
       isModalAlertVariantOpen: !isModalAlertVariantOpen,
-      customEscapePressed: customEscapePressed ?? false 
+      customEscapePressed: customEscapePressed ?? false
     }));
   };
 
@@ -523,7 +523,7 @@ export class ModalNextDemo extends React.Component<React.HTMLProps<HTMLDivElemen
     };
 
     return (
-      <React.Fragment>
+      <>
         <div id="tabstop-test" tabIndex={0} style={{ display: 'flex', flexDirection: 'row', flexWrap: 'wrap' }}>
           <Button style={buttonStyle} variant="primary" onClick={this.handleModalToggle} id="showDefaultModalButton">
             Show Modal
@@ -616,7 +616,7 @@ export class ModalNextDemo extends React.Component<React.HTMLProps<HTMLDivElemen
         {this.renderModalWithAlertVariant()}
         {this.renderHelpModal()}
         {this.renderCustomFocusModal()}
-      </React.Fragment>
+      </>
     );
   }
 }