+++ packages/react-integration/demo-app-ts/src/components/demos/SwitchDemo/SwitchDemo.tsx
@@ -36,7 +36,7 @@ export class SwitchDemo extends React.Component<{}, SwitchState> {
     this.setState({
       table: {
         ...this.state.table,
-        data: this.state.table.data.map(d => (d.id === id ? { ...d, isChecked } : d))
+        data: this.state.table.data.map((d) => (d.id === id ? { ...d, isChecked } : d))
       }
     });
   };
@@ -50,7 +50,7 @@ export class SwitchDemo extends React.Component<{}, SwitchState> {
       simple: { isChecked }
     } = this.state;
     return (
-      <React.Fragment>
+      <>
         <Switch
           id="simple-switch"
           label={<div style={{ color: 'green' }}>Message when on</div>}
@@ -67,7 +67,7 @@ export class SwitchDemo extends React.Component<{}, SwitchState> {
           isChecked={false}
           isDisabled
         />
-      </React.Fragment>
+      </>
     );
   }
 
@@ -82,7 +82,7 @@ export class SwitchDemo extends React.Component<{}, SwitchState> {
       const [_first, ...rest] = data;
       rr = rest;
     }
-    const rows = rr.map(r => ({
+    const rows = rr.map((r) => ({
       cells: [
         {
           title: <Switch id={r.id} isChecked={r.isChecked} onChange={this.handleChangeTable(r.id)} />
@@ -104,10 +104,10 @@ export class SwitchDemo extends React.Component<{}, SwitchState> {
 
   render() {
     return (
-      <React.Fragment>
+      <>
         {this.renderSimple()}
         {this.renderTable()}
-      </React.Fragment>
+      </>
     );
   }
 }