+++ packages/react-integration/demo-app-ts/src/components/demos/TableDemo/TableEditableCompoundExpandableDemo.tsx
@@ -1,10 +1,5 @@
 import * as React from 'react';
-import {
-  compoundExpand,
-  IRow,
-  ICell,
-  IRowCell
-} from '@patternfly/react-table';
+import { compoundExpand, IRow, ICell, IRowCell } from '@patternfly/react-table';
 import { Table, TableHeader, TableBody, TableProps } from '@patternfly/react-table/deprecated';
 
 import CodeBranchIcon from '@patternfly/react-icons/dist/esm/icons/code-branch-icon';
@@ -47,25 +42,25 @@ export class TableEditableCompoundExpandableDemo extends React.Component<TablePr
             { title: <a href="#">siemur/test-space</a>, props: { component: 'th' } },
             {
               title: (
-                <React.Fragment>
+                <>
                   <CodeBranchIcon key="icon" /> 10
-                </React.Fragment>
+                </>
               ),
               props: { isOpen: true, ariaControls: 'compoound-expansion-table-1' }
             },
             {
               title: (
-                <React.Fragment>
+                <>
                   <CodeIcon key="icon" /> 4
-                </React.Fragment>
+                </>
               ),
               props: { isOpen: false, ariaControls: 'compoound-expansion-table-2' }
             },
             {
               title: (
-                <React.Fragment>
+                <>
                   <CubeIcon key="icon" /> 4
-                </React.Fragment>
+                </>
               ),
               props: { isOpen: false, ariaControls: 'compoound-expansion-table-3' }
             },
@@ -124,25 +119,25 @@ export class TableEditableCompoundExpandableDemo extends React.Component<TablePr
             { title: <a href="#">siemur/test-space</a>, props: { component: 'th' } },
             {
               title: (
-                <React.Fragment>
+                <>
                   <CodeBranchIcon key="icon" /> 3
-                </React.Fragment>
+                </>
               ),
               props: { isOpen: false, ariaControls: 'compoound-expansion-table-4' }
             },
             {
               title: (
-                <React.Fragment>
+                <>
                   <CodeIcon key="icon" /> 4
-                </React.Fragment>
+                </>
               ),
               props: { isOpen: false, ariaControls: 'compoound-expansion-table-5' }
             },
             {
               title: (
-                <React.Fragment>
+                <>
                   <CubeIcon key="icon" /> 2
-                </React.Fragment>
+                </>
               ),
               props: { isOpen: false, ariaControls: 'compoound-expansion-table-6' }
             },
@@ -202,7 +197,7 @@ export class TableEditableCompoundExpandableDemo extends React.Component<TablePr
 
   onExpand(_event: React.MouseEvent, rowIndex: number, colIndex: number, isOpen: boolean) {
     const newRows = Array.from(this.state.rows);
-    const rowCells = Array.from(newRows[rowIndex].cells as IRowCell[]) ;
+    const rowCells = Array.from(newRows[rowIndex].cells as IRowCell[]);
     const thisCell = rowCells[colIndex];
 
     if (!isOpen) {
@@ -217,7 +212,7 @@ export class TableEditableCompoundExpandableDemo extends React.Component<TablePr
       rowCells[colIndex] = { ...thisCell, props: { ...thisCell.props, isOpen: false } };
       newRows[rowIndex] = {
         ...newRows[rowIndex],
-        isOpen: rowCells.some(cell => cell.props && cell.props.isOpen)
+        isOpen: rowCells.some((cell) => cell.props && cell.props.isOpen)
       };
     }
     this.setState({