+++ packages/react-integration/demo-app-ts/src/components/demos/TabsDemo/TabsDemo.tsx
@@ -58,7 +58,7 @@ export class TabDemo extends Component {
     const { isTab2Hidden } = this.state;
 
     return (
-      <React.Fragment>
+      <>
         <Button id="showTab2" onClick={() => this.setState({ isTab2Hidden: !isTab2Hidden })}>
           {isTab2Hidden ? 'Show' : 'Hide'} tab 2
         </Button>
@@ -102,7 +102,7 @@ export class TabDemo extends Component {
             ref={this.contentRef1}
             aria-label="Tab item 1"
             // eslint-disable-next-line no-console
-            onAuxClick={event => console.log(event)}
+            onAuxClick={(event) => console.log(event)}
           >
             Tab 1 section
           </TabContent>
@@ -185,7 +185,7 @@ export class TabDemo extends Component {
             Tab 3 section
           </Tab>
         </Tabs>
-      </React.Fragment>
+      </>
     );
   }
 }