+++ packages/react-table/src/demos/examples/TableColumnManagement.tsx
@@ -225,7 +225,7 @@ export const TableColumnManagement: React.FunctionComponent = () => {
         <Button key="save" variant="primary" onClick={onSave}>
           Save
         </Button>,
-        <Button key="cancel" variant="secondary" onClick={handleModalToggle}>
+        <Button key="cancel" variant="link" onClick={handleModalToggle}>
           Cancel
         </Button>
       ]}