+++ packages/react-tokens/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [5.4.0-prerelease.5](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-tokens@5.4.0-prerelease.4...@patternfly/react-tokens@5.4.0-prerelease.5) (2024-05-31)
+
+**Note:** Version bump only for package @patternfly/react-tokens
+
 # [5.4.0-prerelease.4](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-tokens@5.4.0-prerelease.3...@patternfly/react-tokens@5.4.0-prerelease.4) (2024-05-20)
 
 **Note:** Version bump only for package @patternfly/react-tokens